Improved AWS and Azure discoverer
This PR improves the discoverers, mainly for AWS and Azure.
