+++ service/discovery/azure/compute.go
@@ -75,6 +75,13 @@ func (d *azureComputeDiscovery) List() (list []voc.IsCloudResource, err error) {
 		return nil, fmt.Errorf("%s: %w", ErrCouldNotAuthenticate, err)
 	}
 
+	// Discover block storage
+	storage, err := d.discoverBlockStorages()
+	if err != nil {
+		return nil, fmt.Errorf("could not discover block storage: %w", err)
+	}
+	list = append(list, storage...)
+
 	log.Info("Discover Azure compute resources")
 
 	// Discover virtual machines
@@ -167,25 +174,24 @@ func (d *azureComputeDiscovery) discoverVirtualMachines() ([]voc.IsCloudResource
 
 	// List all VMs across all resource groups
 	listPager := client.NewListAllPager(&armcompute.VirtualMachinesClientListAllOptions{})
-	vms := make([]*armcompute.VirtualMachine, 0)
+
 	for listPager.More() {
 		pageResponse, err := listPager.NextPage(context.TODO())
 		if err != nil {
 			err = fmt.Errorf("%s: %v", ErrGettingNextPage, err)
 			return nil, err
 		}
-		vms = append(vms, pageResponse.Value...)
-	}
 
-	for i := range vms {
-		r, err := d.handleVirtualMachines(vms[i])
-		if err != nil {
-			return nil, fmt.Errorf("could not handle virtual machine: %w", err)
+		for _, vm := range pageResponse.Value {
+			r, err := d.handleVirtualMachines(vm)
+			if err != nil {
+				return nil, fmt.Errorf("could not handle virtual machine: %w", err)
+			}
+
+			log.Infof("Adding virtual machine '%s'", r.GetName())
+
+			list = append(list, r)
 		}
-
-		log.Infof("Adding virtual machine %+v", r)
-
-		list = append(list, r)
 	}
 
 	return list, err
@@ -282,4 +288,133 @@ func bootLogOutput(vm *armcompute.VirtualMachine) string {
 		return ""
 	}
 	return ""
+}
+
+func (d *azureComputeDiscovery) discoverBlockStorages() ([]voc.IsCloudResource, error) {
+	var list []voc.IsCloudResource
+
+	// Create disks client
+	client, err := armcompute.NewDisksClient(util.Deref(d.sub.SubscriptionID), d.cred, &d.clientOptions)
+	if err != nil {
+		err = fmt.Errorf("could not get new disks client: %w", err)
+		return nil, err
+	}
+
+	// List all disks across all resource groups
+	listPager := client.NewListPager(&armcompute.DisksClientListOptions{})
+	for listPager.More() {
+		pageResponse, err := listPager.NextPage(context.TODO())
+		if err != nil {
+			err = fmt.Errorf("%s: %w", ErrGettingNextPage, err)
+			return nil, err
+		}
+
+		for _, disk := range pageResponse.Value {
+			blockStorages, err := d.handleBlockStorage(disk)
+			if err != nil {
+				return nil, fmt.Errorf("could not handle block storage: %w", err)
+			}
+			log.Infof("Adding block storage '%s'", blockStorages.Name)
+
+			list = append(list, blockStorages)
+		}
+	}
+
+	return list, nil
+}
+
+func (d *azureComputeDiscovery) handleBlockStorage(disk *armcompute.Disk) (*voc.BlockStorage, error) {
+	// If a mandatory field is empty, the whole disk is empty
+	if disk == nil || disk.ID == nil {
+		return nil, fmt.Errorf("disk is nil")
+	}
+
+	enc, err := d.blockStorageAtRestEncryption(disk)
+	if err != nil {
+		return nil, fmt.Errorf("could not get block storage properties for the atRestEncryption: %w", err)
+	}
+
+	return &voc.BlockStorage{
+		Storage: &voc.Storage{
+			Resource: &voc.Resource{
+				ID:           voc.ResourceID(util.Deref(disk.ID)),
+				ServiceID:    discovery.DefaultCloudServiceID,
+				Name:         util.Deref(disk.Name),
+				CreationTime: disk.Properties.TimeCreated.Unix(),
+				Type:         []string{"BlockStorage", "Storage", "Resource"},
+				GeoLocation: voc.GeoLocation{
+					Region: util.Deref(disk.Location),
+				},
+				Labels: labels(disk.Tags),
+			},
+			AtRestEncryption: enc,
+		},
+	}, nil
+}
+
+// blockStorageAtRestEncryption takes encryption properties of an armcompute.Disk and converts it into our respective
+// ontology object.
+func (d *azureComputeDiscovery) blockStorageAtRestEncryption(disk *armcompute.Disk) (enc voc.IsAtRestEncryption, err error) {
+	var (
+		diskEncryptionSetID string
+		keyUrl              string
+	)
+
+	if disk == nil {
+		return enc, errors.New("disk is empty")
+	}
+
+	if disk.Properties.Encryption.Type == nil {
+		return enc, errors.New("error getting atRestEncryption properties of blockStorage")
+	} else if *disk.Properties.Encryption.Type == armcompute.EncryptionTypeEncryptionAtRestWithPlatformKey {
+		enc = &voc.ManagedKeyEncryption{AtRestEncryption: &voc.AtRestEncryption{
+			Algorithm: "AES256",
+			Enabled:   true,
+		}}
+	} else if *disk.Properties.Encryption.Type == armcompute.EncryptionTypeEncryptionAtRestWithCustomerKey {
+		diskEncryptionSetID = util.Deref(disk.Properties.Encryption.DiskEncryptionSetID)
+
+		keyUrl, err = d.keyURL(diskEncryptionSetID)
+		if err != nil {
+			return nil, fmt.Errorf("could not get keyVaultID: %w", err)
+		}
+
+		enc = &voc.CustomerKeyEncryption{
+			AtRestEncryption: &voc.AtRestEncryption{
+				Algorithm: "", // TODO(garuppel): TBD
+				Enabled:   true,
+			},
+			KeyUrl: keyUrl,
+		}
+	}
+
+	return enc, nil
+}
+
+func (d *azureComputeDiscovery) keyURL(diskEncryptionSetID string) (string, error) {
+	if diskEncryptionSetID == "" {
+		return "", ErrMissingDiskEncryptionSetID
+	}
+
+	// Create Key Vault client
+	client, err := armcompute.NewDiskEncryptionSetsClient(util.Deref(d.sub.SubscriptionID), d.cred, &d.clientOptions)
+	if err != nil {
+		err = fmt.Errorf("could not get new key vault client: %w", err)
+		return "", err
+	}
+
+	// Get disk encryption set
+	kv, err := client.Get(context.TODO(), resourceGroupName(diskEncryptionSetID), diskEncryptionSetName(diskEncryptionSetID), &armcompute.DiskEncryptionSetsClientGetOptions{})
+	if err != nil {
+		err = fmt.Errorf("could not get key vault: %w", err)
+		return "", err
+	}
+
+	keyURL := kv.DiskEncryptionSet.Properties.ActiveKey.KeyURL
+
+	if keyURL == nil {
+		return "", fmt.Errorf("could not get keyURL")
+	}
+
+	return util.Deref(keyURL), nil
 }
